Iteration 9

+++ src/grt/src/NEWGR/include/DataType.h
@@ -253,7 +253,7 @@ struct OrderTree
 {
   int length;
   int treeIndex;
-  int16_t xmin;
+  double xmin;
 };
 
 struct OrderNetEdge
